x updated control for the camera

+++ Phi/Game/Editor/ImGuiEntity.cpp
@@ -4,35 +4,35 @@
 
 namespace Editor
 {
-    // GLFW callbacks to get events
-    static void glfw_error_callback(int error, const char* description)
-    {
-        fputs(description, stderr);
-    }
-    
-    
-    static void glfw_scroll_callback(GLFWwindow* window, double xoffset, double yoffset)
-    {
-        ImGuiIO& io = ImGui::GetIO();
-        io.MouseWheel += (float)yoffset; // Use fractional mouse wheel, 1.0 unit 5 lines.
-    }
-    
-    static void glfw_key_callback(GLFWwindow* window, int key, int scancode, int action, int mods)
-    {
-        ImGuiIO& io = ImGui::GetIO();
-        if (action == GLFW_PRESS)
-            io.KeysDown[key] = true;
-        if (action == GLFW_RELEASE)
-            io.KeysDown[key] = false;
-        io.KeyCtrl = (mods & GLFW_MOD_CONTROL) != 0;
-        io.KeyShift = (mods & GLFW_MOD_SHIFT) != 0;
-    }
-    
-    static void glfw_char_callback(GLFWwindow* window, unsigned int c)
-    {
-        if (c > 0 && c < 0x10000)
-            ImGui::GetIO().AddInputCharacter((unsigned short)c);
-    }
+	// GLFW callbacks to get events
+	static void glfw_error_callback(int error, const char* description)
+	{
+		fputs(description, stderr);
+	}
+	
+	
+	static void glfw_scroll_callback(GLFWwindow* window, double xoffset, double yoffset)
+	{
+		ImGuiIO& io = ImGui::GetIO();
+		io.MouseWheel += (float)yoffset; // Use fractional mouse wheel, 1.0 unit 5 lines.
+	}
+	
+	static void glfw_key_callback(GLFWwindow* window, int key, int scancode, int action, int mods)
+	{
+		ImGuiIO& io = ImGui::GetIO();
+		if (action == GLFW_PRESS)
+			io.KeysDown[key] = true;
+		if (action == GLFW_RELEASE)
+			io.KeysDown[key] = false;
+		io.KeyCtrl = (mods & GLFW_MOD_CONTROL) != 0;
+		io.KeyShift = (mods & GLFW_MOD_SHIFT) != 0;
+	}
+	
+	static void glfw_char_callback(GLFWwindow* window, unsigned int c)
+	{
+		if (c > 0 && c < 0x10000)
+			ImGui::GetIO().AddInputCharacter((unsigned short)c);
+	}
 
 	void ImGuiEntity::RenderDrawLists(ImDrawList** const cmd_lists, int cmd_lists_count)
 	{
@@ -70,7 +70,7 @@ namespace Editor
 			total_vtx_count += cmd_lists[n]->vtx_buffer.size();
 
 		Engine::VertexBuffer::Binder     bind1(entity->mVertexBuffer);
-		size_t needed_buffer_size = total_vtx_count * sizeof(ImDrawVert);
+		int needed_buffer_size = (int) total_vtx_count * sizeof(ImDrawVert);
 		if (needed_buffer_size > entity->mVertexBuffer.size())
 			entity->mVertexBuffer.update(NULL, 0, needed_buffer_size);
 
@@ -103,7 +103,7 @@ namespace Editor
 		}
 		
 		// Restore modified state
-		glScissor(0, 0, width, height);
+		glScissor(0, 0, (int) width, (int) height);
 		glDisable(GL_SCISSOR_TEST);
 		glDisable(GL_TEXTURE_2D);
 	}
@@ -112,32 +112,32 @@ namespace Editor
 	{
 		int width, height;
 		glfwGetFramebufferSize(initialize.window, &width, &height);
-        glfwSetCharCallback(initialize.window, glfw_char_callback);
-        glfwSetKeyCallback(initialize.window, glfw_key_callback);
-        glfwSetScrollCallback(initialize.window, glfw_scroll_callback);
-        
+		glfwSetCharCallback(initialize.window, glfw_char_callback);
+		glfwSetKeyCallback(initialize.window, glfw_key_callback);
+		glfwSetScrollCallback(initialize.window, glfw_scroll_callback);
+		
 		ImGuiIO& io = ImGui::GetIO();
 		io.DisplaySize = ImVec2((float)width, (float)height);  // Display size, in pixels. For clamping windows positions.
 		io.PixelCenterOffset = 0.0f;                                  // Align OpenGL texels
 		io.RenderDrawListsFn = RenderDrawLists;
 		io.UserData = (void*)this;
-        io.KeyMap[ImGuiKey_Tab] = GLFW_KEY_TAB;                       // Keyboard mapping. ImGui will use those indices to peek into the io.KeyDown[] array.
-        io.KeyMap[ImGuiKey_LeftArrow] = GLFW_KEY_LEFT;
-        io.KeyMap[ImGuiKey_RightArrow] = GLFW_KEY_RIGHT;
-        io.KeyMap[ImGuiKey_UpArrow] = GLFW_KEY_UP;
-        io.KeyMap[ImGuiKey_DownArrow] = GLFW_KEY_DOWN;
-        io.KeyMap[ImGuiKey_Home] = GLFW_KEY_HOME;
-        io.KeyMap[ImGuiKey_End] = GLFW_KEY_END;
-        io.KeyMap[ImGuiKey_Delete] = GLFW_KEY_DELETE;
-        io.KeyMap[ImGuiKey_Backspace] = GLFW_KEY_BACKSPACE;
-        io.KeyMap[ImGuiKey_Enter] = GLFW_KEY_ENTER;
-        io.KeyMap[ImGuiKey_Escape] = GLFW_KEY_ESCAPE;
-        io.KeyMap[ImGuiKey_A] = GLFW_KEY_A;
-        io.KeyMap[ImGuiKey_C] = GLFW_KEY_C;
-        io.KeyMap[ImGuiKey_V] = GLFW_KEY_V;
-        io.KeyMap[ImGuiKey_X] = GLFW_KEY_X;
-        io.KeyMap[ImGuiKey_Y] = GLFW_KEY_Y;
-        io.KeyMap[ImGuiKey_Z] = GLFW_KEY_Z;
+		io.KeyMap[ImGuiKey_Tab] = GLFW_KEY_TAB;                       // Keyboard mapping. ImGui will use those indices to peek into the io.KeyDown[] array.
+		io.KeyMap[ImGuiKey_LeftArrow] = GLFW_KEY_LEFT;
+		io.KeyMap[ImGuiKey_RightArrow] = GLFW_KEY_RIGHT;
+		io.KeyMap[ImGuiKey_UpArrow] = GLFW_KEY_UP;
+		io.KeyMap[ImGuiKey_DownArrow] = GLFW_KEY_DOWN;
+		io.KeyMap[ImGuiKey_Home] = GLFW_KEY_HOME;
+		io.KeyMap[ImGuiKey_End] = GLFW_KEY_END;
+		io.KeyMap[ImGuiKey_Delete] = GLFW_KEY_DELETE;
+		io.KeyMap[ImGuiKey_Backspace] = GLFW_KEY_BACKSPACE;
+		io.KeyMap[ImGuiKey_Enter] = GLFW_KEY_ENTER;
+		io.KeyMap[ImGuiKey_Escape] = GLFW_KEY_ESCAPE;
+		io.KeyMap[ImGuiKey_A] = GLFW_KEY_A;
+		io.KeyMap[ImGuiKey_C] = GLFW_KEY_C;
+		io.KeyMap[ImGuiKey_V] = GLFW_KEY_V;
+		io.KeyMap[ImGuiKey_X] = GLFW_KEY_X;
+		io.KeyMap[ImGuiKey_Y] = GLFW_KEY_Y;
+		io.KeyMap[ImGuiKey_Z] = GLFW_KEY_Z;
 
 		const void* png_data;
 		unsigned int png_size;
@@ -196,7 +196,7 @@ namespace Editor
 		io.DeltaTime = update.delta;
 
 		ImGui::NewFrame();
-        //ImGui::SetWindowFontScale(2.f);
+		//ImGui::SetWindowFontScale(2.f);
 }
 	
 	void ImGuiEntity::destroy(Game::Destroy& destroy)
